requirements.txt
alembic==1.0.11
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.0
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.1.0
markdown2==2.3.8
MarkupSafe==1.1.1
pep8==1.7.1
Pillow==6.1.0
psycopg2==2.8.3
pycparser==2.19
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.6
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
